add webrtc proj and remove boost deps rename std::shared_ptr<DataPacket> packetPtr;

+++ erizo/src/test/stats/MovingAverageStatTest.cpp
@@ -51,37 +51,37 @@ TEST_F(MovingAverageStatTest, shouldCalculateAverageForLessSamplesThanWindowSize
 }
 
 TEST_F(MovingAverageStatTest, shouldCalculateAverageForValuesInWindow) {
-  for (uint i = 0; i < kArbitraryWindowSize; i++) {
+  for (int i = 0; i < kArbitraryWindowSize; i++) {
     moving_average_stat+=kArbitraryNumberToAdd;
   }
-  for (uint i = 0; i < kArbitraryWindowSize; i++) {
+  for (int i = 0; i < kArbitraryWindowSize; i++) {
     moving_average_stat+=kArbitraryNumberToAdd + 1;
   }
   EXPECT_EQ(moving_average_stat.value(), kArbitraryNumberToAdd + 1);
 }
 
 TEST_F(MovingAverageStatTest, shouldCalculateAverageForValuesInWindow2) {
-  for (uint i = 0; i < kArbitraryWindowSize; i++) {
+  for (int i = 0; i < kArbitraryWindowSize; i++) {
     moving_average_stat+=kArbitraryNumberToAdd;
   }
-  for (uint i = 0; i < kArbitraryWindowSize; i++) {
+  for (int i = 0; i < kArbitraryWindowSize; i++) {
     moving_average_stat+=kArbitraryNumberToAdd - 1;
   }
   EXPECT_EQ(moving_average_stat.value(), kArbitraryNumberToAdd - 1);
 }
 
 TEST_F(MovingAverageStatTest, shouldCalculateAverageForGivenNumberOfSamples) {
-  for (uint i = 0; i < kArbitraryWindowSize; i++) {
+  for (int i = 0; i < kArbitraryWindowSize; i++) {
     moving_average_stat+=i;
   }
   EXPECT_EQ(moving_average_stat.value(1), kArbitraryWindowSize - 1);
-  EXPECT_EQ(moving_average_stat.value(2), uint((kArbitraryWindowSize * 2 - 3) / 2));
+  EXPECT_EQ(moving_average_stat.value(2), ((kArbitraryWindowSize * 2 - 3) / 2));
 }
 
 TEST_F(MovingAverageStatTest, shouldCalculateAverageForGivenNumberOfSamples2) {
-  for (uint i = kArbitraryWindowSize; i > 0; i--) {
+  for (int i = kArbitraryWindowSize; i > 0; i--) {
     moving_average_stat+=i;
   }
   EXPECT_EQ(moving_average_stat.value(1), 1u);
-  EXPECT_EQ(moving_average_stat.value(2), uint((1 + 2) / 2));
+  EXPECT_EQ(moving_average_stat.value(2), ((1 + 2) / 2));
 }
